Bump package version
I had a weird bug while publishing and I figured out that the only way
to avoid it was to downgrade the version of the bower NPM dependency.
I have no clue why this was necessary, but since the fix in fc5f3fa is
super important, I decided that the sort term solution of having a
working downgraded version of bower was acceptable.